+++ CHANGELOG.md
@@ -3,7 +3,7 @@
     Placeholder for the next version (at the beginning of the line):
     ## **WORK IN PROGRESS**
 -->
-## **WORK IN PROGRESS**
+## 3.1.0 (2021-09-15)
 * New plugin `license` to check for outdated license years
 * New plugin `manual-review` to do manual changes before committing
 * Fix: `git` plugin never `amends` commits